% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.urlscan}
\alias{print.urlscan}
\title{Print for urlscan objects}
\usage{
\method{print}{urlscan}(x, ...)
}
\arguments{
\item{x}{urlscan object}

\item{...}{ignored}
}
\description{
Print for urlscan objects
}
\keyword{internal}
