% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.urlscan_result}
\alias{print.urlscan_result}
\title{Print for urlscan_result objects}
\usage{
\method{print}{urlscan_result}(x, ...)
}
\arguments{
\item{x}{urlscan_result object}

\item{...}{ignored}
}
\description{
Print for urlscan_result objects
}
\keyword{internal}
