% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.urlscan_submit}
\alias{print.urlscan_submit}
\title{Print for urlscan_submit objects}
\usage{
\method{print}{urlscan_submit}(x, ...)
}
\arguments{
\item{x}{urlscan_submit object}

\item{...}{ignored}
}
\description{
Print for urlscan_submit objects
}
\keyword{internal}
