% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlscan-package.R
\docType{package}
\name{urlscan}
\alias{urlscan}
\alias{urlscan-package}
\title{Analyze Websites and Resources They Request}
\description{
The <urlscan.io> service provides an 'API' enabling analysis of
websites and the resources they request. Much like the 'Inspector' of your
browser, <urlscan.io> will let you take a look at the individual resources
that are requested when a site is loaded. Tools are provided to search
public <urlscans.io> scan submissions.
}
\author{
Bob Rudis (bob@rud.is)
}
