% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{urlscan_result}
\alias{urlscan_result}
\title{Retrieve detailed results for a given scan ID}
\usage{
urlscan_result(scan_id, include_dom = FALSE, include_shot = FALSE)
}
\arguments{
\item{scan_id}{scan id (UUID)}

\item{include_dom}{(logical) include the website DOM? (default: \code{FALSE})}

\item{include_shot}{(logical) include the website screen shot? (default: \code{FALSE})}
}
\value{
\code{list} with \code{scan_result} task, page, content lists, fetch data,
connection metadata and computed stats.\cr
\cr
The list can also include \code{dom} if
\code{include_dom} is \code{TRUE}. If so, \code{dom} will be an \code{httr} \code{response} object
since the data could be binary. Use \code{httr} tools to process it.\cr
\cr
The list can also include \code{screenshot} if \code{include_shot} is \code{TRUE} and
a screenshot was available.
}
\description{
Retrieve detailed results for a given scan ID
}
\examples{
x <- urlscan_search("domain:r-project.org")
r <- urlscan_result(x$results[["_id"]][1])
}
