% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{urlscan_submit}
\alias{urlscan_submit}
\title{Submit a URL for scanning}
\usage{
urlscan_submit(url, public = TRUE, custom_agent = NULL,
  referer = NULL, api_key = urlscan_api_key())
}
\arguments{
\item{url}{URL to submit}

\item{public}{Public or private results? \code{TRUE} (default) = public.}

\item{custom_agent}{(character) Override User-Agent for this scan}

\item{referer}{(character) Override HTTP referer for this scan}

\item{api_key}{your API key. See \code{\link[=urlscan_api_key]{urlscan_api_key()}}}
}
\value{
The response to the API call will give you the ID and API endpoint for t
he scan, you can use it to retrieve the result after waiting for a short while.
}
\description{
urlscan offers API access to submit URLs for researchers and other institutions.
Using the API requires an API key. If you'd like to start submitting URLs via the
API, contact them at \url{mailto:info@urlscan.io} with a short description on how
you're planning to use the API. Submissions via the API can be public or private,
though they prefer public submissions so other users can benefit from the results
of your scans.\cr
\cr
The API key should be in the environment variable \code{URLSCAN_API_KEY} and the
easiest way to do that is via the \code{~/.Renviron} file. You can also manually pass
it in as a parameter.
}
\examples{
\dontrun{
urlscan_submit("https://r-project.org/")
}
}
\references{
\url{https://urlscan.io/about-api/#submission}
}
